auspice only displays local files
Part of a move to a future where
auspice.us handles community datasets & narratives,
nextstrain.org handles S3 datasets

Github pages builds are currently not working as
(e.g.) `/dist/bundle.js` should be `dist/bundle.js`
But this change would break server builds.
